ai-conversational-python-backend-repo/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg[binary]==3.2.13
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-dotenv==1.1.0
pydantic-settings==2.9.1
dateparser==1.2.1
alembic==1.15.2
SQLAlchemy==2.0.40
email-validator==2.2.0
